Changed name of the map file.
Added exceptions to importing the map function.
